Renamed package model.domain to model.entity
