+++ src/main/java/com/retail/dvdapplication/model/entity/DVD.java
@@ -1,4 +1,4 @@
-package com.retail.dvdapplication.model.domain;
+package com.retail.dvdapplication.model.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.PositiveOrZero;
