+++ src/main/java/com/retail/dvdapplication/model/entity/Employee.java
@@ -1,4 +1,4 @@
-package com.retail.dvdapplication.model.domain;
+package com.retail.dvdapplication.model.entity;
 
 import jakarta.persistence.*;
