+++ src/main/java/com/retail/dvdapplication/model/repository/EmployeeRepository.java
@@ -1,6 +1,6 @@
 package com.retail.dvdapplication.model.repository;
 
-import com.retail.dvdapplication.model.domain.Employee;
+import com.retail.dvdapplication.model.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /*
